+++ CHANGELOG.md
@@ -6,6 +6,13 @@ Patch releases are only for test purposes - here I only document major and minor
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.7.0] - 2025-06-27
+
+### Added
+
+- content slider (image slider)
+- buttons active state
+
 ## [2.6.0] - 2025-06-25
 
 ### Added